+++ lib/dynamic_material.dart
@@ -1,4 +1,3 @@
-import 'morphable_shape_border.dart';
 import 'package:flutter/material.dart';
 
 ///A simpler version of the Material class