+++ CHANGELOG.md
@@ -13,4 +13,8 @@ Fix control point drag lag when editing freeform shapes
 
 ## [0.0.4] - 02.04.2021
 
-Lets you select morph method 
+Lets you select morph method
+
+## [0.0.5] - 02.05.2021
+
+Fixed some static code warnings